+++ widgets/tws_widgets/lib/src/core/models/tws_state_holder.dart
@@ -1,0 +1,4 @@
+import 'package:csm_view/csm_view.dart';
+
+/// [TWSFStateHolder] Basic [CSMStateBase] class declaration.
+final class TWSFStateHolder extends CSMStateBase { }